fix: split up component
